migrated code from tuner.jsx to the body.jsx and implemented use effect to make the preset buttons responsive.
